Error off because we never casted to int

+++ testapp/beaglelogic_seniorDesignV2.c
@@ -201,7 +201,7 @@ int main(int argc, char **argv)
 
 				/* Check past with present values */
 				if (buffer[i] != buffer[i-2] || buffer[i + 1] != buffer[i-1]){
-					changeState(buffer[i], buffer[i + 1]);
+					changeState((int) buffer[i], (int) buffer[i + 1]);
 				}
 				/* check to see if we need to transmit */
 				if (pub_signal){
